# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dar Al Maghrib Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: Get reviews (public)
      parameters:
        - name: menuItemId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # COUPONS
...
components:
  schemas:
...
    Review:
      type: object
      required: [id, userId, menuItemId, rating, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        menuItemId:
          type: integer
        menuItemName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [menuItemId, rating]
      properties:
        menuItemId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # FAVORITES
...